requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.2
argparse==1.2.1
django-autoslug==1.7.1
django-crispy-forms==1.4.0
django-debug-toolbar==0.9.4
django-endless-pagination==2.0
django-geoposition==0.1.4
django-jfu==1.0.7
django-model-utils==1.5.0
django-notifications-hq==0.6.0
django-smart-selects==1.0.9
pgmagick==0.5.7
psycopg2==2.5.1
python-memcached==1.53
sorl-thumbnail==11.12
wsgiref==0.1.2
yolk==0.4.3
